requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
PyJWT[crypto]